requirements.txt
transformers==4.23.1
datasets==2.5.2
sentencepiece==0.1.97
torch==1.12.1
streamlit==1.8.1
matplotlib
Pillow
